+++ hs/test/HelVM/HelMA/Automata/Cat/EvaluatorSpec.hs
@@ -3,7 +3,7 @@ module HelVM.HelMA.Automata.Cat.EvaluatorSpec (spec) where
 import           HelVM.HelMA.Automata.Cat.Evaluator
 
 import           HelVM.HelMA.Automaton.API.IOTypes
-import           HelVM.HelMA.Automaton.Eff.MockEff
+import           HelVM.HelMA.Automaton.Eff.Mock
 
 import           Test.Hspec                         (Spec, describe, it)
 import           Test.Hspec.Expectations.Pretty