+++ hs/test/HelVM/HelMA/Automata/WhiteSpace/EvaluatorSpec.hs
@@ -5,7 +5,7 @@ import           HelVM.HelMA.Automata.WhiteSpace.Evaluator
 import           HelVM.HelMA.Automata.WhiteSpace.FileExtra
 import           HelVM.HelMA.Automata.WhiteSpace.SimpleParams
 
-import           HelVM.HelMA.Automaton.Eff.MockEff
+import           HelVM.HelMA.Automaton.Eff.Mock
 
 import           HelVM.HelMA.Automaton.Types.LabelType